// app/lib/types/experience.ts
export interface Experience {
  id: number;
  cover: string;
  title: string;
  subtitle: string;
  description: string;
  rex: string;
  mission: string;
  type: string;
  duration: string;
  website: string;
  skills: string[];
  screenShot: string;
  screenShot1: string;
  screenShot2: string;
  isMobile: boolean;
}